Various bug fixes (#234)
* Implemented name for contract type. (TDB if super case is properly handled)

* Bug fix in type checking for exponentiation.

* Bug fix: Added type conversion in implementation of member "name" for contracts.

* Bug fix: Removed type conversion of exponent to result type in exponentiation.

* Updated failing tests: syntaxTests/{conversion, enums, indexing, literalOperations, lvalues, metaTypes}, semanticTests/{cleanup, conversions, integer, literals, metaTypes}

* Updated failing compilation tests.

+++ test/libsolidity/syntaxTests/metaTypes/integer.sol
@@ -1,12 +1,12 @@
 contract Test {
     function basic() public pure {
-        uint uintMax = type(uint).max;
+        uint256 uintMax = type(uint256).max;
         uintMax;
-        int intMax = type(int).max;
+        int256 intMax = type(int256).max;
         intMax;
-        uint uintMin = type(uint).min;
+        uint256 uintMin = type(uint256).min;
         uintMin;
-        int intMin = type(int).min;
+        int256 intMin = type(int256).min;
         intMin;
     }
 }
