+++ test/libsolidity/syntaxTests/metaTypes/integer_pure.sol
@@ -1,16 +1,16 @@
 contract test {
 
     function viewAssignment() public view {
-        int min = type(int).min;
+        int256 min = type(int256).min;
         min;
     }
 
     function assignment() public {
-        int max = type(int).max;
+        int256 max = type(int256).max;
         max;
     }
 
 }
 // ----
-// Warning 2018: (21-112): Function state mutability can be restricted to pure
-// Warning 2018: (118-200): Function state mutability can be restricted to pure
+// Warning 2018: (21-118): Function state mutability can be restricted to pure
+// Warning 2018: (124-212): Function state mutability can be restricted to pure
